+++ CHANGELOG.md
@@ -100,6 +100,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - chore: add SPDX license sections
 - refactor(src): replace useColor ternaries with helper function calls
 - docs(install): add instructions trying vendr with nix
+- refactor(src): replace ternaries in vendr::log functions, move verbose/useColor out of log ns
 
 ### Added
 - build(packaging): added '-git' variant of PKGBUILD which follows repository HEAD